+++ database/migrations/2014_10_11_000001_create_currencies_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('name')->unique();
             $table->string('country_name')->nullable();
             $table->string('code', 10)->unique();
-            $table->double('rate', 10, 2)->nullable();
+            $table->decimal('rate', 10, 2)->nullable();
             $table->string('symbol', 10)->nullable();
             $table->string('position', 20)->nullable();
             $table->boolean('status')->default(1);
