+++ database/migrations/2023_12_26_161841_create_purchases_table.php
@@ -17,11 +17,11 @@ return new class extends Migration
             $table->foreignId('business_id')->constrained()->cascadeOnDelete();
             $table->foreignId('user_id')->nullable()->constrained()->nullOnDelete();
             $table->foreignId('tax_id')->nullable()->constrained()->nullOnDelete();
-            $table->double('discountAmount', 10, 2)->default(0);
-            $table->double('tax_amount', 10, 2)->default(0);
-            $table->double('dueAmount', 10, 2)->default(0);
-            $table->double('paidAmount', 10, 2)->default(0);
-            $table->double('totalAmount', 10, 2)->default(0);
+            $table->decimal('discountAmount', 10, 2)->default(0);
+            $table->decimal('tax_amount', 10, 2)->default(0);
+            $table->decimal('dueAmount', 10, 2)->default(0);
+            $table->decimal('paidAmount', 10, 2)->default(0);
+            $table->decimal('totalAmount', 10, 2)->default(0);
             $table->string('invoiceNumber')->nullable();
             $table->boolean('isPaid')->default(0);
             $table->string('paymentType')->default('Cash');
