+++ database/migrations/2023_12_26_163232_create_purchase_details_table.php
@@ -15,11 +15,11 @@ return new class extends Migration
             $table->id();
             $table->foreignId('purchase_id')->constrained()->cascadeOnDelete();
             $table->foreignId('product_id')->constrained()->cascadeOnDelete();
-            $table->double('purchase_without_tax')->default(0);
-            $table->double('purchase_with_tax')->default(0);
-            $table->double('profit_percent')->default(0);
-            $table->double('sales_price')->default(0);
-            $table->double('wholesale_price')->default(0);
+            $table->decimal('purchase_without_tax')->default(0);
+            $table->decimal('purchase_with_tax')->default(0);
+            $table->decimal('profit_percent')->default(0);
+            $table->decimal('sales_price')->default(0);
+            $table->decimal('wholesale_price')->default(0);
             $table->integer('quantities')->default(0);
             $table->string('batch_no')->nullable();
             $table->date('expire_date')->nullable();
