+++ database/migrations/2023_12_27_131216_create_due_collects_table.php
@@ -19,9 +19,9 @@ return new class extends Migration
             $table->foreignId('sale_id')->nullable()->constrained()->nullOnDelete();
             $table->foreignId('purchase_id')->nullable()->constrained()->cascadeOnDelete();
             $table->string('invoiceNumber')->nullable();
-            $table->double('totalDue', 10, 2)->default(0);
-            $table->double('dueAmountAfterPay', 10, 2)->default(0);
-            $table->double('payDueAmount', 10, 2)->default(0);
+            $table->decimal('totalDue', 10, 2)->default(0);
+            $table->decimal('dueAmountAfterPay', 10, 2)->default(0);
+            $table->decimal('payDueAmount', 10, 2)->default(0);
             $table->string('paymentType')->default('Cash');
             $table->timestamp('paymentDate')->nullable();
             $table->timestamps();
